requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --resolver=backtracking
#
click==8.1.3
    # via
    #   derivativeworksblog (setup.py)
    #   flask
flask==2.2.3
    # via derivativeworksblog (setup.py)
gunicorn==20.1.0
    # via derivativeworksblog (setup.py)
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markdown2==2.4.8
    # via derivativeworksblog (setup.py)
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
pyyaml==6.0
    # via derivativeworksblog (setup.py)
werkzeug==2.2.3
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
